pyproject.toml
[project]
name = "surfer-mcp"
version = "0.1.0"
description = "MCP server for controlling Surfer waveform viewer via WCP"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyobjc-framework-quartz>=12.1; sys_platform == 'darwin'",
    "pywellen>=0.20.0",
]

[project.scripts]
surfer-mcp = "surfer_mcp.server:main"
simvision-mcp = "simvision_mcp.server:main"

[project.optional-dependencies]
dev = ["pytest>=7"]
screenshot = [
    "mss>=10.2.0",
    "pillow>=12.2.0",
    "python-xlib>=0.33",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/surfer_mcp", "src/simvision_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
